pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "upwork-proposal-agent"
version = "1.0.0"
description = "Gemini-powered Upwork proposal generator"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-generativeai>=0.3.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
